+++ mpush-core/src/main/java/com/mpush/core/ack/AckTask.java
@@ -29,16 +29,16 @@ import java.util.concurrent.Future;
  * @author ohun@live.cn (夜色)
  */
 public final class AckTask implements Runnable {
-    private final int ackMessageId;
+    private final long ackMessageId;
     private AckTaskQueue ackTaskQueue;
     private AckCallback callback;
     private Future<?> timeoutFuture;
 
-    public AckTask(int ackMessageId) {
+    public AckTask(long ackMessageId) {
         this.ackMessageId = ackMessageId;
     }
 
-    public static AckTask from(int ackMessageId) {
+    public static AckTask from(long ackMessageId) {
         return new AckTask(ackMessageId);
     }
 
@@ -56,10 +56,9 @@ public final class AckTask implements Runnable {
         return this;
     }
 
-    public int getAckMessageId() {
+    public long getAckMessageId() {
         return ackMessageId;
     }
-
 
     private boolean tryDone() {
         return timeoutFuture.cancel(true);
